requirements.txt
alembic==1.5.8
aniso8601==9.0.1
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
openpyxl==3.0.7
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.4.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
retrying==1.3.3
six==1.15.0
SQLAlchemy==1.4.9
Werkzeug==1.0.1
